requirements.txt
streamlit
langchain-community
langchain-core
google-generativeai
python-dotenv
pdf2image
